get rid of lying, useless error message in login_write()
also remove incorrect comments about the return value of login_login() and
login_logout(); fwiw, it's actually the reverse -- 0 for success and 1
for failure.

and while at it, better make 'void' the return type of functions whose
return value is always ignored.

also remove leftover prototype of non-existing login_log_entry() function.

--------------------------------------

when running sshd as a regular user, as with:

  ssh-keygen -qN '' -f ./srv-key
  /usr/sbin/sshd -o PidFile=none -f /dev/null -De -p 9797 -h $(pwd)/srv-key

you get a dubious message like:

   Attempt to write login records by non-root user (aborting)

everytime a connection is accepted.

But the server is NOT aborting the connection (and in fact, it does not
care at all if the login records were written or not) and no attempt to
do anything untoward happened.